+++ PointOfSale.Models/InvoiceDetail.cs
@@ -20,7 +20,7 @@ namespace PointOfSale.Models
         public int ProductId { get; set; }
         [ForeignKey(nameof(ProductId))]
         public Product Product { get; set; }
-        public int Count {  get; set; }
+        public double Count {  get; set; }
 
         public double Price {  get; set; }
